requirements.txt
Pillow==6.2.2
torch==1.3.1
torchvision==0.4.2
opencv-python==4.1.2.30
scikit-image==0.16.2
scipy==1.3.3
lmdb==0.98
colour==0.1.5
numpy=1.18.1
